Azure AI Multi-Agent Ticket Triage System/requirements.txt
azure-ai-projects
azure-ai-agents
azure-identity
python-dotenv
pydantic